requirements.txt
# requirements.txt
pydantic
requests
fastapi
uvicorn
pydantic-ai
openai
python-dotenv
